chore: move usermap api to custom project
