+++ src/Christofel.sln
@@ -34,6 +34,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Discord.Net.WebSocket", "..
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Christofel.Api", "Christofel.Api\Christofel.Api.csproj", "{27156F3A-DDEB-4178-85F4-50D0F5FB4446}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Apis", "Apis", "{0B39810A-5D62-42DD-A20F-FF100A95A0E8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Usermap", "Usermap\Usermap.csproj", "{72C4AF34-C5BA-41EF-8782-49D88C1C4413}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,6 +56,7 @@ Global
 		{8D950F70-39D9-4EBB-832E-3FE683DAA694} = {B7C5F5E3-A588-41F9-AC21-7465030396F6}
 		{F6E41005-7D33-48F8-9705-287447260338} = {B7C5F5E3-A588-41F9-AC21-7465030396F6}
 		{27156F3A-DDEB-4178-85F4-50D0F5FB4446} = {05930E47-11FB-41BE-BEBB-5CD09C29117C}
+		{72C4AF34-C5BA-41EF-8782-49D88C1C4413} = {0B39810A-5D62-42DD-A20F-FF100A95A0E8}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0C798625-3BD0-40A1-8921-08292D0445A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -114,5 +119,9 @@ Global
 		{27156F3A-DDEB-4178-85F4-50D0F5FB4446}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{27156F3A-DDEB-4178-85F4-50D0F5FB4446}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{27156F3A-DDEB-4178-85F4-50D0F5FB4446}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72C4AF34-C5BA-41EF-8782-49D88C1C4413}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72C4AF34-C5BA-41EF-8782-49D88C1C4413}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72C4AF34-C5BA-41EF-8782-49D88C1C4413}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72C4AF34-C5BA-41EF-8782-49D88C1C4413}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
