+++ src/Usermap/Usermap.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="5.0.0" />
+      <PackageReference Include="Microsoft.Extensions.Options" Version="5.0.0" />
+      <PackageReference Include="RestSharp" Version="106.12.0" />
+      <PackageReference Include="RestSharp.Serializers.NewtonsoftJson" Version="106.12.0" />
+    </ItemGroup>
+
+</Project>
